se modificaron las llaves primarias y tambien se agrego nombre a las llaves forign key

+++ src/main/java/com/registro/control_acceso/repository/LectorRepository.java
@@ -4,6 +4,6 @@ import com.registro.control_acceso.entity.Lector;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LectorRepository extends JpaRepository<Lector, Long> {
+public interface LectorRepository extends JpaRepository<Lector, Integer> {
     Optional<Lector> findByTarjetaRfid(String tarjetarfid);
 }
