+++ src/main/java/com/registro/control_acceso/repository/PuertaRepository.java
@@ -3,5 +3,5 @@ package com.registro.control_acceso.repository;
 import com.registro.control_acceso.entity.Puerta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PuertaRepository extends JpaRepository<Puerta, Long> {
+public interface PuertaRepository extends JpaRepository<Puerta, Integer> {
 }
